OAUTH2-231 Remove duplicate code, make it available for everybody
