+++ modules/apps/oauth2-provider/oauth2-provider-web/src/main/java/com/liferay/oauth2/provider/web/internal/portlet/action/GenerateRandomSecretMVCResourceCommand.java
@@ -14,8 +14,8 @@
 
 package com.liferay.oauth2.provider.web.internal.portlet.action;
 
+import com.liferay.oauth2.provider.util.OAuth2SecureRandomGenerator;
 import com.liferay.oauth2.provider.web.internal.constants.OAuth2ProviderPortletKeys;
-import com.liferay.oauth2.provider.web.internal.display.context.OAuth2AdminPortletDisplayContext;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCResourceCommand;
 
 import java.io.IOException;
@@ -51,7 +51,7 @@ public class GenerateRandomSecretMVCResourceCommand
 			PrintWriter printWriter = resourceResponse.getWriter();
 
 			String randomSecret =
-				OAuth2AdminPortletDisplayContext.generateRandomSecret();
+				OAuth2SecureRandomGenerator.generateRandomSecret();
 
 			printWriter.write(randomSecret);
 		}
